Allow Discord linking for untracked streamers
