+++ cogs/twitch/dashboard/__init__.py
@@ -1,0 +1,5 @@
+"""Twitch dashboard package."""
+
+from .app import Dashboard, build_app
+
+__all__ = ["Dashboard", "build_app"]
