+++ cogs/twitch/leaderboard.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
+from datetime import datetime, timezone
 from typing import Any, Dict, List, Optional, Sequence, Tuple
 
 import discord
@@ -448,19 +449,89 @@ class TwitchLeaderboardMixin:
         out: Dict[str, Any] = {"tracked": {}, "category": {}}
 
         tracked_logins: set[str] = set()
+        verified_logins: set[str] = set()
+        discord_lookup: Dict[str, Dict[str, Any]] = {}
+
+        def _parse_db_datetime(value: Any) -> Optional[datetime]:
+            if not value:
+                return None
+            if isinstance(value, datetime):
+                dt = value
+            else:
+                try:
+                    dt = datetime.fromisoformat(str(value))
+                except ValueError:
+                    return None
+            if dt.tzinfo is None:
+                dt = dt.replace(tzinfo=timezone.utc)
+            return dt.astimezone(timezone.utc)
+
+        now_utc = datetime.now(timezone.utc)
         try:
             with storage.get_conn() as c:
                 rows = c.execute(
-                    "SELECT twitch_login FROM twitch_streamers"
+                    """
+                    SELECT twitch_login,
+                           is_on_discord,
+                           discord_user_id,
+                           discord_display_name,
+                           manual_verified_permanent,
+                           manual_verified_until
+                      FROM twitch_streamers
+                    """
                 ).fetchall()
             for row in rows:
-                login_raw = row[0] if isinstance(row, tuple) else row["twitch_login"]
-                login = str(login_raw or "").strip().lower()
-                if login:
-                    tracked_logins.add(login)
+                if isinstance(row, tuple):
+                    columns = (
+                        "twitch_login",
+                        "is_on_discord",
+                        "discord_user_id",
+                        "discord_display_name",
+                        "manual_verified_permanent",
+                        "manual_verified_until",
+                    )
+                    row_data = {
+                        key: row[idx] if idx < len(row) else None for idx, key in enumerate(columns)
+                    }
+                elif hasattr(row, "keys"):
+                    row_data = {key: row[key] for key in row.keys()}
+                else:
+                    row_data = dict(row)
+
+                login = str(row_data.get("twitch_login") or "").strip().lower()
+                if not login:
+                    continue
+
+                tracked_logins.add(login)
+
+                is_on_discord_raw = row_data.get("is_on_discord", 0)
+                discord_id_raw = row_data.get("discord_user_id")
+                discord_name_raw = row_data.get("discord_display_name")
+
+                is_verified = False
+                try:
+                    if row_data.get("manual_verified_permanent"):
+                        is_verified = True
+                    else:
+                        until_dt = _parse_db_datetime(row_data.get("manual_verified_until"))
+                        if until_dt and until_dt >= now_utc:
+                            is_verified = True
+                except Exception:
+                    log.debug("Konnte Verifizierungsstatus für %s nicht lesen", login, exc_info=True)
+
+                if is_verified:
+                    verified_logins.add(login)
+
+                discord_lookup[login] = {
+                    "is_on_discord": int(is_on_discord_raw or 0),
+                    "discord_user_id": discord_id_raw,
+                    "discord_display_name": discord_name_raw,
+                }
         except Exception:
             log.exception("Konnte gespeicherte Twitch-Logins nicht laden")
             tracked_logins = set()
+            verified_logins = set()
+            discord_lookup = {}
 
         def _normalize_hour(value: Optional[int]) -> Optional[int]:
             if value is None:
@@ -538,12 +609,41 @@ class TwitchLeaderboardMixin:
                 return items
             for item in items:
                 login = str(item.get("streamer") or "").strip().lower()
-                if login and login in tracked_logins:
-                    item["is_partner"] = 1
+                if not login:
+                    continue
+                if login in tracked_logins:
+                    item["is_partner"] = 1 if login in verified_logins else 0
             return items
 
-        out["tracked"]["top"] = _apply_partner_flag(_aggregate(tracked_sql, hour_params))
-        out["category"]["top"] = _apply_partner_flag(_aggregate(category_sql, hour_params))
+        def _apply_discord_info(
+            items: List[dict], *, assume_members: Optional[set[str]] = None
+        ) -> List[dict]:
+            assumed_members = assume_members or set()
+            for item in items:
+                login = str(item.get("streamer") or "").strip().lower()
+                info = discord_lookup.get(login, {})
+                discord_user_id = info.get("discord_user_id") if info else None
+                discord_display_name = info.get("discord_display_name") if info else None
+                has_profile = bool(
+                    (discord_user_id and str(discord_user_id).strip())
+                    or (discord_display_name and str(discord_display_name).strip())
+                )
+                default_member = login in assumed_members
+                is_member = default_member or bool(info.get("is_on_discord")) or has_profile
+                item["is_on_discord"] = 1 if is_member else 0
+                item["discord_user_id"] = discord_user_id
+                item["discord_display_name"] = discord_display_name
+                item["has_discord_profile"] = 1 if has_profile else 0
+            return items
+
+        out["tracked"]["top"] = _apply_discord_info(
+            _apply_partner_flag(_aggregate(tracked_sql, hour_params)),
+            assume_members=verified_logins,
+        )
+        out["category"]["top"] = _apply_discord_info(
+            _apply_partner_flag(_aggregate(category_sql, hour_params)),
+            assume_members=verified_logins,
+        )
 
         tracked_hourly_sql = (
             """
