+++ cogs/twitch/storage.py
@@ -104,6 +104,7 @@ def ensure_schema(conn: sqlite3.Connection) -> None:
             next_link_check_at         TEXT,
             discord_user_id            TEXT,
             discord_display_name       TEXT,
+            is_on_discord              INTEGER DEFAULT 0,
             manual_verified_permanent  INTEGER DEFAULT 0,
             manual_verified_until      TEXT,
             manual_verified_at         TEXT,
@@ -117,6 +118,7 @@ def ensure_schema(conn: sqlite3.Connection) -> None:
         ("next_link_check_at", "TEXT"),
         ("discord_user_id", "TEXT"),
         ("discord_display_name", "TEXT"),
+        ("is_on_discord", "INTEGER DEFAULT 0"),
         ("manual_verified_permanent", "INTEGER DEFAULT 0"),
         ("manual_verified_until", "TEXT"),
         ("manual_verified_at", "TEXT"),
